Update README.md
We are excited to announce the development of EasyLoanForm, a user-friendly 
loan application tool designed to simplify and streamline the loan application 
process. We are currently seeking talented front-end developers with expertise in 
HTML, CSS, and JavaScript to contribute to this project. Your contributions will 
help enhance the user experience, improve functionality, and ensure a seamless 
interaction for users. If you are passionate about creating intuitive web 
applications and would like to be a part of this innovative project, we invite you 
to join us and make a meaningful impact on EasyLoanForm.